Created graphs for individual transaction, and a timeline of these transactions across the last 12 months.
Decided to slighly change the approach of this project, and for me to be able to host it
and be reached by other users, i have decided i will treat the login as the up token.

This will lead to a removal of most of the backend, and essentially create a react only app.